# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "api autentia manager",
        "title": "Autentia Manager Documentation",
        "contact": {},
        "version": "1"
    },
    "paths": {
...
        "/logout": {
            "delete": {
                "security": [
                    {
                        "barerToken": []
                    }
                ],
                "description": "user logout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "user logout",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Usuario no activo"
                    }
                }
            }
        },
...
        "/users": {
            "get": {
                "security": [
                    {
                        "barerToken": []
                    }
                ],
                "description": "users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/routes.userStruct"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/routes.MessageResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "barerToken": []
                    }
                ],
                "description": "create user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "user",
                "parameters": [
                    {
                        "description": "user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.userParams"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Usuario Creado"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/routes.MessageResponse"
                        }
                    }
                }
            }
        },
...
        "/users/feedback": {
            "post": {
                "security": [
                    {
                        "barerToken": []
                    }
                ],
                "description": "user feedback",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "user feedback",
                "parameters": [
                    {
                        "description": "feedback",
                        "name": "feedback",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.feedBackParams"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/routes.feedbackResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/routes.MessageResponse"
                        }
                    }
                }
            }
        },
...
        "/users/introspection": {
            "get": {
                "security": [
                    {
                        "barerToken": []
                    }
                ],
                "description": "instrospection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "user instrospection",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.response"
                        }
                    }
                }
            }
        },
...
        "/users/{nickName}": {
            "put": {
                "security": [
                    {
                        "barerToken": []
                    }
                ],
                "description": "update user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "update user",
                "parameters": [
                    {
                        "description": "user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.updateUserParams"
                        }
                    },
                    {
                        "type": "string",
                        "description": "userIdentifier",
                        "name": "nickName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "El usuario se ha actualizado con éxito"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/routes.MessageResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "barerToken": []
                    }
                ],
                "description": "user delete",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "user delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "nickName",
                        "name": "nickName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/routes.MessageResponse"
                        }
                    }
                }
            }
        },
...
        "/users/{userIdentifier}": {
            "get": {
                "security": [
                    {
                        "barerToken": []
                    }
                ],
                "description": "user get",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "user get",
                "parameters": [
                    {
                        "type": "string",
                        "description": "userIdentifier",
                        "name": "userIdentifier",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/routes.MessageResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "data.Country": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "data.User": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "country": {
                    "$ref": "#/definitions/data.Country"
                },
                "dni": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nick_name": {
                    "type": "string"
                },
                "picture": {
                    "type": "string"
                },
                "status_user": {
                    "type": "string"
                },
                "validated": {
                    "type": "boolean"
                }
            }
        },
...
        "routes.MessageResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "string"
                }
            }
        },
...
        "routes.feedBackParams": {
            "type": "object",
            "properties": {
                "browser": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "system": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "routes.feedbackResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "routes.response": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/data.User"
                }
            }
        },
...
        "routes.roleParams": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "routes.updateUserParams": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "dni": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "status_user": {
                    "type": "string"
                }
            }
        },
        "routes.userParams": {
            "type": "object",
            "required": [
                "country",
                "dni",
                "email",
                "name",
                "nickname"
            ],
            "properties": {
                "country": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "dni": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "routes.userStruct": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "country": {
                    "$ref": "#/definitions/data.Country"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nick_name": {
                    "type": "string"
                },
                "picture": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/routes.roleParams"
                    }
                },
                "status_user": {
                    "type": "string"
                },
                "validated": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "barerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}